+++ eclients/utils.py
@@ -6,12 +6,19 @@
 @software: PyCharm
 @time: 18-12-26 下午3:32
 """
+import sys
 from collections import MutableMapping, MutableSequence
 from contextlib import contextmanager
 
-import aelog
+import yaml
+from bson import ObjectId
 
-__all__ = ("ignore_error", "verify_message")
+try:
+    from yaml import CLoader as Loader
+except ImportError:
+    from yaml import Loader
+
+__all__ = ("ignore_error", "verify_message", "analysis_yaml", "gen_class_name", "objectid")
 
 
 @contextmanager
@@ -23,10 +30,11 @@ def ignore_error(error=Exception):
     Returns:
 
     """
+    # noinspection PyBroadException
     try:
         yield
-    except error as e:
-        aelog.warning(e)
+    except error:
+        pass
 
 
 def verify_message(src_message: dict, message: list or dict):
@@ -47,3 +55,42 @@ def verify_message(src_message: dict, message: list or dict):
             if set(msg.keys()).intersection(required_field) == required_field and msg["msg_code"] in src_message:
                 src_message[msg["msg_code"]].update(msg)
     return src_message
+
+
+def gen_class_name(underline_name):
+    """
+    由下划线的名称变为驼峰的名称
+    Args:
+        underline_name
+    Returns:
+
+    """
+    return "".join([name.capitalize() for name in underline_name.split("_")])
+
+
+def analysis_yaml(full_conf_path):
+    """
+    解析yaml文件
+    Args:
+        full_conf_path: yaml配置文件路径
+    Returns:
+
+    """
+    with open(full_conf_path, 'rt', encoding="utf8") as f:
+        try:
+            conf = yaml.load(f, Loader=Loader)
+        except yaml.YAMLError as e:
+            print("Yaml配置文件出错, {}".format(e))
+            sys.exit()
+    return conf
+
+
+def objectid():
+    """
+
+    Args:
+
+    Returns:
+
+    """
+    return str(ObjectId())